+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.1.0
+
+* removed deprecated package io/ioutil, the minimum Go version is now 1.16
+
 ## 1.0.1
 
 * fixed selecting largest icon